amber_aim/pyproject.toml
[project]
name = "aim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
package-mode = true

dependencies = [
    "boto3>=1.29.0",
    "fastapi>=0.104.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.24.0",
    "twelvelabs>=0.2.0",
    "ipykernel>=6.30.1",
    "openai>=2.1.0",
    "openai-agents>=0.1.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "black>=23.11.0",
    "httpx>=0.25.0",
    "ipdb>=0.13.13",
    "isort>=5.12.0",
    "moto>=4.2.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.ruff.lint.isort]
known-first-party = ["aim"]
